+++ assets/api/assets/assets.py
@@ -1,10 +1,10 @@
-from typing import List
-
 from ninja import NinjaAPI
+
+from django.db.models import Sum
 
 from assets.api import schema
 from assets.hooks import get_extension_logger
-from assets.models import Assets
+from assets.models import Assets, Request, RequestAssets
 
 logger = get_extension_logger(__name__)
 
@@ -14,29 +14,42 @@ class AssetsApiEndpoints:
 
     def __init__(self, api: NinjaAPI):
         @api.get(
-            "assets/filter/{location}/",
-            response={200: List[schema.Assets], 403: str},
+            "assets/location/{location_id}/{location_flag}/",
+            response={200: schema.Assets, 403: str},
             tags=self.tags,
             auth=None,
         )
-        def get_assets(request, location: str):
+        def get_assets(request, location_id: int, location_flag: str):
             perms = request.user.has_perm("assets.basic_access")
 
             if not perms:
                 return 403, "Permission Denied"
 
-            assets = Assets.objects.all().select_related("location", "eve_type")
-            assets = assets.filter(location_flag=location, location_id=1042478386825)
+            assets_qs = Assets.objects.all().select_related("location", "eve_type")
+            assets_qs = assets_qs.filter(
+                location_flag=location_flag, location_id=location_id
+            )
 
-            output = []
+            assets = []
 
-            for asset in assets:
+            for asset in assets_qs:
+                ordered_quantity = RequestAssets.objects.filter(
+                    asset=asset,
+                    requestor__status=Request.STATUS_OPEN,
+                ).aggregate(total_quantity=Sum("quantity"))["total_quantity"]
+
+                if ordered_quantity:
+                    if ordered_quantity >= asset.quantity:
+                        continue
+                    asset.quantity -= ordered_quantity
+
                 try:
                     price = asset.price * asset.quantity
                 except TypeError:
                     price = "N/A"
-                output.append(
+                assets.append(
                     {
+                        "asset_pk": asset.pk,
                         "item_id": asset.item_id,
                         "name": asset.eve_type.name,
                         "quantity": asset.quantity,
@@ -49,5 +62,11 @@ class AssetsApiEndpoints:
                         "price": price,
                     }
                 )
+
+            output = {
+                "location_id": location_id,
+                "location_flag": location_flag,
+                "assets": assets,
+            }
 
             return output