+++ assets/api/requests/requests.py
@@ -1,14 +1,16 @@
-from typing import List, Any
+from typing import Any, List
 
 from ninja import NinjaAPI
 
-from django.utils import timezone
 from django.core.handlers.wsgi import WSGIRequest
+from django.shortcuts import render
+from django.utils import timezone
+from django.utils.translation import gettext_lazy as _
 
 from assets.api import schema
+from assets.api.requests.helper import _request_actions, _request_list
 from assets.hooks import get_extension_logger
-from assets.models import Request
-from assets.api.requests.helper import _request_actions
+from assets.models import Request, RequestAssets
 
 logger = get_extension_logger(__name__)
 
@@ -16,7 +18,7 @@ logger = get_extension_logger(__name__)
 class RequestsApiEndpoints:
     tags = ["Requests"]
 
-    def __init__(self, api: NinjaAPI):       
+    def __init__(self, api: NinjaAPI):
         @api.get(
             "requests/",
             response={200: List[schema.Requests], 403: str},
@@ -37,6 +39,10 @@ class RequestsApiEndpoints:
                 status=Request.STATUS_CANCELLED, closed_at__lt=skip_old_entrys
             )
 
+            requests_data = requests_data.exclude(
+                status=Request.STATUS_COMPLETED, closed_at__lt=skip_old_entrys
+            )
+
             output = []
 
             for req in requests_data:
@@ -44,7 +50,11 @@ class RequestsApiEndpoints:
                     {
                         "id": req.pk,
                         "status": req.get_status_display(),
-                        "order": "",
+                        "order": _request_list(
+                            req,
+                            perms,
+                            request,
+                        ),
                         "action": req.status,
                         "created": req.created_at,
                         "closed": req.closed_at,
@@ -89,7 +99,11 @@ class RequestsApiEndpoints:
                     {
                         "id": req.pk,
                         "status": req.get_status_display(),
-                        "order": "",
+                        "order": _request_list(
+                            req,
+                            perms,
+                            request,
+                        ),
                         "action": req.status,
                         "created": req.created_at,
                         "closed": req.closed_at,
@@ -124,10 +138,56 @@ class RequestsApiEndpoints:
                 requests_count = None
 
             my_requests_count = Request.objects.my_requests_total_count(request.user)
-            
+
             output = {
                 "requestCount": requests_count,
                 "myRequestCount": my_requests_count,
             }
 
             return output
+
+        @api.get(
+            "requests/order/{request_id}/",
+            response={200: Any, 403: str},
+            tags=self.tags,
+        )
+        def get_request_order(request: WSGIRequest, request_id: int):
+            """Get the order for a request"""
+            perms = request.user.has_perm("assets.basic_access")
+
+            if not perms:
+                return 403, "Permission Denied"
+
+            try:
+                request_user = Request.objects.get(
+                    pk=request_id,
+                )
+            except Request.DoesNotExist:
+                return 403, "Request does not exist"
+
+            orders = RequestAssets.objects.filter(
+                requestor__requesting_user=request.user,
+                requestor__pk=request_id,
+            )
+
+            output = []
+
+            for order in orders:
+                output.append(
+                    {
+                        "item_id": order.asset.eve_type.id,
+                        "name": order.asset.eve_type.name,
+                        "quantity": order.quantity,
+                    }
+                )
+
+            context = {
+                "request_id": request_id,
+                "orders": output,
+                "title": _("Order for {requestor} - ID: {request_id}").format(
+                    requestor=request_user.requesting_user.username,
+                    request_id=request_id,
+                ),
+            }
+
+            return render(request, "assets/partials/modal/view-order.html", context)