+++ assets/static/assets/js/modals/single-request.js
@@ -1,7 +1,8 @@
 $(document).ready(() => {
     /* global tableAssets, loadRequestStatistics */
     const modalRequestOrder = $('#assets-single-request');
-    const modalErrorMessage = $('#modal-error-message');
+    const modalErrorMessage = modalRequestOrder.find('#modal-error-message');
+    const modalErrorRequiredText = modalErrorMessage.text();
 
     // Approve Request Modal
     modalRequestOrder.on('show.bs.modal', (event) => {
@@ -18,14 +19,31 @@ $(document).ready(() => {
         const modalDiv = modalRequestOrder.find('#modal-request-text');
         modalDiv.html(modalText);
 
+        const form = modalRequestOrder.find('form');
+        const amountField = form.find('input[name="amount"]');
+        const itemQuantity = button.data('item-quantity');
+
+        // Set the max attribute for the amount field
+        amountField.attr('max', itemQuantity);
+
+        // Automatically correct the input value if it exceeds the max value
+        amountField.on('input', () => {
+            const currentValue = parseInt(amountField.val(), 10);
+            if (currentValue > itemQuantity) {
+                amountField.val(itemQuantity); // Set to max value
+            }
+        });
+
+        $('#modal-button-buy-single-request').on('click', () => {
+            amountField.val(itemQuantity); // Set the amount to the maximum quantity
+        });
+
         $('#modal-button-confirm-single-request').on('click', () => {
-            const form = modalRequestOrder.find('form');
-            const amountField = form.find('input[name="amount"]');
             const amount = amountField.val();
             const csrfMiddlewareToken = form.find('input[name="csrfmiddlewaretoken"]').val();
 
-            if (!amount) {
-                modalErrorMessage.removeClass('d-none');
+            if (!amount || amount > itemQuantity) {
+                modalErrorMessage.text(modalErrorRequiredText).removeClass('d-none');
                 amountField.addClass('is-invalid');
 
                 // Add shake class to the error field
@@ -40,8 +58,7 @@ $(document).ready(() => {
                     url,
                     {
                         amount: amount,
-                        item_id: button.data('item-id'),
-                        location_id: button.data('location-id'),
+                        asset_pk: button.data('asset-pk'),
                         csrfmiddlewaretoken: csrfMiddlewareToken
                     }
                 );
@@ -50,15 +67,20 @@ $(document).ready(() => {
                     modalRequestOrder.modal('hide');
                     loadRequestStatistics(); // Reload the request statistics
                     const tableAssets = $('#assets').DataTable();
+                    const tableMyRequest = $('#my-requests').DataTable();
+                    const tableRequest = $('#requests').DataTable();
                     tableAssets.ajax.reload(); // Reload the assets table
+                    tableMyRequest.ajax.reload(); // Reload the my requests table
+                    tableRequest.ajax.reload(); // Reload the requests table
                 }).fail((xhr, _, __) => {
                     const response = JSON.parse(xhr.responseText);
                     modalErrorMessage.text(response.message).removeClass('d-none'); // Show the error message
-                    modalErrorMessage.addClass('l-shake'); // Add the shake class
+                    // Add shake class to the error field
+                    modalErrorMessage.addClass('ts-shake');
 
-                    // Remove the shake class after 2 seconds
+                    // Remove the shake class after 3 seconds
                     setTimeout(() => {
-                        modalErrorMessage.removeClass('l-shake');
+                        modalErrorMessage.removeClass('ts-shake');
                     }, 2000);
                 });
             }