+++ CHANGELOG.md
@@ -8,7 +8,11 @@
 - Modal System
 - Confirm Modal
 - Single Request Modal
+- View Order Modal
+- Multi Request Modal
 - Request Statistic API
+- Button System
+- RequestAssets Model
 
 ### Changed